category/games/2026/02/20260209-043051-vector-frogger-river-cross/pyproject.toml
[project]
name = "vector-frogger-river-cross"
version = "1.0.0"
description = "Navigate a frog across a busy river by jumping on moving logs"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.scripts]
vector-frogger-river-cross = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
